Fix handling of compressed sass
This should fix the handling of compressed sass. Previously the capture would grab everything from 'content: "' to the last '"' in a line. This was particularly problematic when using outputStyle: 'compressed' which results in most (or all) of the file being on one line.

It also captures the whitespace between 'content:' and '"', and restores the exact whitespace used in the original file. Previously it always wrote out 'content: "' with one space. This could result in unnecessary extra space being added to an already compressed file.

I've added a second test for that runs sass with outputStyle: 'compressed', as well as a file called font-awesome-compressed.css to test against.